+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayerPlugin.java
@@ -92,7 +92,8 @@ public class VideoPlayerPlugin implements FlutterPlugin, AndroidVideoPlayerApi {
             flutterState.applicationContext,
             VideoPlayerEventCallbacks.bindTo(flutterState.binaryMessenger, streamInstance),
             videoAsset,
-            sharedOptions);
+            sharedOptions,
+            options.getForwardBufferDurationMs());
 
     registerPlayerInstance(videoPlayer, id);
     return id;
@@ -112,7 +113,8 @@ public class VideoPlayerPlugin implements FlutterPlugin, AndroidVideoPlayerApi {
             VideoPlayerEventCallbacks.bindTo(flutterState.binaryMessenger, streamInstance),
             handle,
             videoAsset,
-            sharedOptions);
+            sharedOptions,
+            options.getForwardBufferDurationMs());
 
     registerPlayerInstance(videoPlayer, id);
     return new TexturePlayerIds(id, handle.id());