+++ packages/video_player/video_player_android/android/src/test/java/io/flutter/plugins/videoplayer/VideoPlayerPluginTest.java
@@ -73,7 +73,7 @@ public class VideoPlayerPluginTest {
     try (MockedStatic<PlatformViewVideoPlayer> mockedPlatformViewVideoPlayerStatic =
         mockStatic(PlatformViewVideoPlayer.class)) {
       mockedPlatformViewVideoPlayerStatic
-          .when(() -> PlatformViewVideoPlayer.create(any(), any(), any(), any()))
+          .when(() -> PlatformViewVideoPlayer.create(any(), any(), any(), any(), any()))
           .thenReturn(mock(PlatformViewVideoPlayer.class));
 
       final CreationOptions options =
@@ -81,6 +81,7 @@ public class VideoPlayerPluginTest {
               "https://flutter.github.io/assets-for-api-docs/assets/videos/bee.mp4",
               null,
               new HashMap<>(),
+              null,
               null);
 
       final long playerId = plugin.createForPlatformView(options);
@@ -95,7 +96,7 @@ public class VideoPlayerPluginTest {
     try (MockedStatic<TextureVideoPlayer> mockedTextureVideoPlayerStatic =
         mockStatic(TextureVideoPlayer.class)) {
       mockedTextureVideoPlayerStatic
-          .when(() -> TextureVideoPlayer.create(any(), any(), any(), any(), any()))
+          .when(() -> TextureVideoPlayer.create(any(), any(), any(), any(), any(), any()))
           .thenReturn(mock(TextureVideoPlayer.class));
 
       final CreationOptions options =
@@ -103,6 +104,7 @@ public class VideoPlayerPluginTest {
               "https://flutter.github.io/assets-for-api-docs/assets/videos/bee.mp4",
               null,
               new HashMap<>(),
+              null,
               null);
 
       final TexturePlayerIds ids = plugin.createForTextureView(options);