+++ packages/video_player/video_player_avfoundation/darwin/RunnerTests/TestClasses.swift
@@ -74,6 +74,7 @@ final class TestAsset: NSObject, FVPAVAsset {
 final class StubPlayerItem: NSObject, FVPAVPlayerItem {
   let asset: FVPAVAsset
   var videoComposition: AVVideoComposition?
+  var preferredForwardBufferDuration: TimeInterval = 0
 
   init(asset: FVPAVAsset = TestAsset()) {
     self.asset = asset