+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation_objc/include/video_player_avfoundation_objc/FVPAVFactory.h
@@ -62,6 +62,9 @@ NS_ASSUME_NONNULL_BEGIN
 
 /// Wraps the underlying playerItem's videoComposition property.
 @property(nonatomic, copy, nullable) AVVideoComposition *videoComposition;
+
+/// Wraps the underlying playerItem's preferredForwardBufferDuration property.
+@property(nonatomic) NSTimeInterval preferredForwardBufferDuration;
 @end
 
 #if TARGET_OS_IOS